New pages added
I have added 3 pages the dashboard, schedule and also analytics(not complete I don't really know how to do the graphs part)
