+++ src/Pages/AnalyticsSummary.css
@@ -1,0 +1,122 @@
+.analytics-summary {
+    padding: 20px;
+    font-family: 'Segoe UI', sans-serif;
+    background: #f8faff;
+}
+
+.top-cards {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+    margin-bottom: 30px;
+}
+
+.card-1 {
+    background: white;
+    padding: 15px;
+    width: 140px;
+    border-radius: 12px;
+    border: 1px solid #ccc;
+}
+.card2 {
+    background: white;
+    padding: 15px;
+    width: 140px;
+    border-radius: 12px;
+    border: 1px solid #ccc;
+}
+
+.small-card {
+    background: white;
+    padding: 15px;
+    width: 140px;
+    border-radius: 12px;
+    border: 1px solid #ccc;
+}
+
+.card .title,
+.small-card .title {
+    font-size: 14px;
+    color: #666;
+}
+
+.value-with-icon {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    font-size: 22px;
+    font-weight: bold;
+}
+
+.card .subtext {
+    font-size: 12px;
+    color: gray;
+    margin-top: 4px;
+}
+
+.export-btn {
+    height: 40px;
+    padding: 0 50px;
+    background: #2563eb;
+    color: white;
+    border: none;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+.bottom-charts {
+    display: flex;
+    gap: 30px;
+}
+
+.line-chart {
+    flex: 2;
+    background: white;
+    padding: 10px;
+    border-radius: 12px;
+    box-shadow: 0 0 6px rgba(0,0,0,0.05);
+    position: relative;
+}
+
+.months {
+    display: flex;
+    justify-content: space-between;
+    padding-top: 10px;
+    font-size: 12px;
+    color: gray;
+}
+
+.donut-card {
+    flex: 1;
+    background: white;
+    padding: 20px;
+    border-radius: 12px;
+    box-shadow: 0 0 6px rgba(0,0,0,0.05);
+    text-align: center;
+}
+
+.donut-chart {
+    position: relative;
+    width: 120px;
+    height: 120px;
+    margin: 0 auto 20px;
+}
+
+.ring {
+    position: absolute;
+    border-radius: 50%;
+    width: 100%;
+    height: 100%;
+    border: 10px solid transparent;
+}
+
+.nbt {
+    border-top-color: #1e3a8a;
+    transform: rotate(0deg);
+}
+
+.math {
+    border-top-color: #6366f1;
+    transform: rotate(120deg);
+}
+
