+++ src/Pages/Dashboard.css
@@ -1,0 +1,118 @@
+body {
+    font-family: 'Segoe UI', sans-serif;
+    margin: 0;
+    padding: 20px;
+    background-color: #F8FAFC;
+}
+
+.dashboard-container {
+    max-width: 1200px;
+    margin: auto;
+    background-color: white;
+    padding: 30px;
+    border-radius: 8px;
+}
+
+.dashboard-actions {
+    display: flex;
+    gap: 16px;
+    margin: 20px 0;
+    flex-wrap: wrap;
+}
+
+.action-card {
+    flex: 1;
+    min-width: 150px;
+    background-color: #fff;
+    border: 1px solid #e5e5e5;
+    text-align: center;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 1px 3px rgba(0,0,0,0.05);
+    font-size: 14px;
+    color: #333;
+}
+
+.main-content {
+    display: flex;
+    gap: 20px;
+    flex-wrap: wrap;
+}
+
+.sessions {
+    flex: 2;
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+}
+
+.session-card {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 15px;
+    padding: 15px;
+    border-radius: 8px;
+    border: 1px solid #e0e0e0;
+    background-color: #f8f8f8;
+}
+
+.in-progress {
+    background-color: #e9f9ee;
+}
+
+.session-meta {
+    font-size: 13px;
+    color: #777;
+}
+
+.join-now {
+    background-color: #28a745;
+    color: white;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+.prepare-btn {
+    background-color: #0d6efd;
+    color: white;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+.snapshot {
+    flex: 1;
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+}
+
+.snapshot-grid {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 12px;
+    margin-top: 10px;
+}
+
+.snap-box {
+    background-color: #f8f9fa;
+    border-radius: 8px;
+    padding: 15px;
+    font-size: 14px;
+    color: #333;
+}
+
+.snap-box strong {
+    font-size: 18px;
+    display: block;
+    margin: 5px 0;
+}
+
+.up {
+    color: green;
+    font-size: 12px;
+}
