+++ src/Pages/studirectory.css
@@ -1,0 +1,152 @@
+/* ==============================================
+   BASE STYLES 
+   ============================================== */
+   * {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+  }
+  
+  body {
+    font-family: system-ui, -apple-system, sans-serif;
+  }
+  
+  /* ==============================================
+     MAIN CONTAINER & NAVBAR 
+     ============================================== */
+  .app-container {
+    width: 100%;
+    min-height: 100vh;
+    background-color: transparent;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .navbar {
+    width: 100%;
+    height: 62px;
+    background: linear-gradient(to right, #0078A6, #5DE0E6);
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    position: relative;
+    padding: 0 20px;
+    display: flex;
+    align-items: center;
+  }
+  
+  .navbar-logo {
+    height: 55px;
+    position: absolute;
+    top: 4px;
+    left: 20px;
+  }
+  
+  .navbar-icons {
+    display: flex;
+    gap: 20px;
+    position: absolute;
+    right: 20px;
+    height: 100%;
+    align-items: center;
+  }
+  
+  .navbar-icon {
+    width: 24px;
+    height: 24px;
+    color: white;
+    cursor: pointer;
+  }
+  
+  /* ==============================================
+     SIDEBAR (MOBILE/DESKTOP) 
+     ============================================== */
+  .sidebar {
+    width: 250px;
+    background-color: #2c3e50;
+    color: white;
+    display: flex;
+    flex-direction: column;
+    position: fixed;
+    top: 0;
+    right: -250px;
+    height: 100vh;
+    transition: right 0.3s ease-in-out;
+    z-index: 1000;
+  }
+  
+  .sidebar.open {
+    right: 0;
+  }
+  
+  .sidebar-header {
+    padding: 15px 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #34495e;
+  }
+  
+  .sidebar-nav {
+    list-style: none;
+    padding: 20px 0;
+  }
+  
+  .sidebar-nav a {
+    display: block;
+    padding: 10px 20px;
+    color: white;
+    text-decoration: none;
+  }
+  
+  /* ==============================================
+     STUDENT CARDS & FILTERS 
+     ============================================== */
+  .data-rectangle {
+    width: 100%;
+    max-width: 400px;
+    min-height: 250px;
+    background: #FFFFFF;
+    border: 1px solid #F1F2F5;
+    border-radius: 8px;
+    padding: 20px 15px 15px 80px;
+    position: relative;
+  }
+  
+  .student-avatar {
+    position: absolute;
+    top: 20px;
+    left: 20px;
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+  }
+  
+  .progress-bar {
+    height: 10px;
+    width: 100%;
+    background-color: #F1F2F5;
+    border-radius: 2px;
+  }
+  
+  .progress-fill {
+    height: 100%;
+    background: #FF2116;
+    width: 75%;
+  }
+  
+  /* ==============================================
+     RESPONSIVE BREAKPOINTS 
+     ============================================== */
+  @media (min-width: 1025px) {
+    /* Desktop styles */
+    .sidebar {
+      position: relative;
+      right: 0;
+    }
+  }
+  
+  @media (max-width: 1024px) {
+    /* Mobile styles */
+    .hamburger-menu {
+      display: flex;
+    }
+  }
